minor bug on time_sequence_visualization

+++ middle_process.py
@@ -780,7 +780,7 @@ def time_sequence_visualization(start_frame, n_frames_showed, end_frame, data, t
     session_name = header['path_session'].split('/')[-2]+'-'+header['path_session'].split('/')[-3].split('-')[1]
     # Array with indeces of considered frames: it starts from the last considerd zero_frames
     considered_frames = np.round(np.linspace(start_frame-1, end_frame-1, n_frames_showed))
-    pieces = np.ceil(len(data)/max_trials)
+    pieces = int(np.ceil(len(data)/max_trials))
     tmp_list = list()
     separators = np.linspace(0, len(data), pieces, endpoint=True)
     # Implementation for splitting big matrices for storing
